Add POST request form to /new_recipe
Previously ingredients where being inserted in a textarea element.
To be able to pass the ingredients as data to the backend, that
element has been replaced with a form that will make a POST request.
The request hasn't benn handled yet in the backend.
Now, ingredients are being stored inside the form, and a submit button
is inserted after the first ingredient is saved.

Some aesthetic changes were also made:
  - Navbar background color.
  - Login and register buttons colors.