Complete new_recipe POST request
POST request has been completed to send recipe data to backend.

If user is anonymous, the data is saved in the session object. This
function seems to be complete, but further test is needed.

If user is logged in, data will be saved in database, but this has not
been yet implemented.

Database model right now needs every ingredient and recipe to be
unique. This will be fixed on the next commit. The idea is that the
database should have an entry for every ingredient, which will also
have a column with the name of the recipe. That way, every ingredient
can be fetched with the recipe's name:

Example:
  Recipe's name | Ingredient  | Other | Columns | User id
  Sandwich      |   Bread     | Other | Columns |    1
  Sandwich      |   Lettuce   | Other | Columns |    1
  Sandwich      |   Soy Ham   | Other | Columns |    1
  Sandwich      |   Tomatoe   | Other | Columns |    1
  Sandwich      |   Bread     | Other | Columns |    2
  Sandwich      |   Onion     | Other | Columns |    2
  Sandwich      |   Beef      | Other | Columns |    2
  Sandwich      |   Tomatoe   | Other | Columns |    2