Fix database structure to hold repeat ingredients
Database had a mistake on the ingredients table, where the recipe and
ingredient's names had to be unique in their respective column.

The way the database is designed, the Ingredient table now holds a row
for every ingredient, and a column of said row contains the recipe's
name.
This way the database can be queryed to create a dictionary with
key : value pairs where the key is the recipe's name, and the value is
a list that contains all the required ingredients.

In addition, the backend now can retrieve all recipes of the current
user and pass them to the recipes.html template, which will show them
to the user.
This has yet to be fully implemented.