Restructure database tables
An Ingredient tables used to hold both the recipes and the ingredients.

This was separated into two different tables: Recipe, and Ingredient.
This will allow to save an image for the recipe on the Recipe table,
while previously every ingredient row would have had an image column.

A new table (RecipePicture) was also added to store the recipe's
images using the sqlalchemy extension "image attach". The logic for
doing thi has not been yet implemented.