Add 'menu maker' function for anonymous users
The framework of this app is practically done
(at least for the basic functionality that I want to implement...
for now)

Changes:
  - An anonymous user now can create menus with their recipes. However
    this info will only last until the session expires.
  - A landing page (responsive) has been created to outline a basic
    guide on how to use this app.
  - Previously there was a problem with the collapsible nav bar, where
    the logged in user icon didn't collapse with the other elements.
    This has been fixed.

There still are a lot of changes and upgrades that I want to add, but
for now this will do as a first draft of the Menu Maker app.

More changes to come!
Hopefully....

Author: Leobardo Arguelles <leobardomed@gmail.com>