Fix ingredient bugs in New Recipe
There were several bugs related to the ingredient box that appears
when adding a new Recipe:
  - The text box couldn't be selected to edit the ingredient
  - You couldn't add a new ingredient simply by pressing enter, you
    had to press the "Add" button

The first bug was fixed to allow for modifications of typos or simple
errors without having to delete the ingredient with the mistake, and
adding a whole new ingredient.
This created a new bug where you can edit the ingredient, but if you
pressed enter, the associated 'delete' button would fire, and it
would erase this ingredient. This has also been fixed.

The second bug was simply to make the process of adding ingredients
easier.